Refactor ModelDTOs for a more consistent folder structure
