+++ src/Services/Ordering/Ordering.FunctionalTests/OrderingScenarios.cs
@@ -3,7 +3,7 @@ using System.Net;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
-using WebMVC.Models;
+using WebMVC.ServicesModelDTOs;
 using Xunit;
 
 namespace Ordering.FunctionalTests
