+++ src/Web/WebMVC/Services/ILocationService.cs
@@ -1,5 +1,5 @@
 ﻿using System.Threading.Tasks;
-using WebMVC.Models;
+using WebMVC.Services.ModelDTOs;
 
 namespace WebMVC.Services
 {
