+++ src/Web/WebMVC/Services/LocationService.cs
@@ -6,7 +6,7 @@ using Newtonsoft.Json;
 using System.Net.Http;
 using System.Threading.Tasks;
 using WebMVC.Infrastructure;
-using WebMVC.Models;
+using WebMVC.Services.ModelDTOs;
 
 namespace WebMVC.Services
 {
