+++ test/ServicesTests/Application.FunctionalTests/Services/Ordering/OrderingScenarios.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
-using WebMVC.Models;
+using WebMVC.Services.ModelDTOs;
 using Xunit;
 
 namespace FunctionalTests.Services.Ordering
